Reprocess complete: 100% annotation fidelity, final main-cell outcomes
- Annotation audit: 19,371 MATCH, 0 gaps/missing/orphans (100.00%)
- Outcome soundness: 100% (all clean/not_cleaned verified vs reports)
- heal_failed cleared to 0 on retry (claude x2, gemini x4)
- FINAL outcomes (clean_at_start/clean/not_cleaned, n=2922 each):
  deepseek 1665/1067/190, qwen 1545/1225/152, glm 1322/1359/241,
  kimi 2035/802/85, gpt 2235/649/38, claude 2276/614/32,
  gemini 2185/701/36
- Data frozen for Results; channel comparison to be refreshed

+++ collected_code_6/claude-fable-5/feedback-static+dynamic-r0-t1.0-p1.0/heal_1/healed/problem-483.c
@@ -16,7 +16,9 @@ static long long first_factorial_divisible_by(long long x)
 
     remainder = 1LL;
     for (n = 1LL; n <= x; n++) {
-        remainder = (remainder * (n % x)) % x;
+        remainder = (long long)(((unsigned __int128)(unsigned long long)remainder *
+                                 (unsigned long long)(n % x)) %
+                                (unsigned long long)x);
         if (remainder == 0LL) {
             return n;
         }
