+++ collected_code_6/gemini-3-pro/feedback-static+dynamic-r0-t1.0-p1.0/heal_10/to_be_healed/problem-919-s3.c
@@ -2,32 +2,35 @@
 #include <stdlib.h>
 
 struct Node {
-    long long data;
+    int data;
     struct Node* next;
 };
  /* Possible weaknesses found:
-  *  test case 0 failed: expected -6, got <no output>
   *  test case 2 failed: expected 18, got <no output>
   *  test case 1 failed: expected 24, got <no output>
+  *  test case 0 failed: expected -6, got <no output>
   */
 
 long long multiplyListItems(struct Node* head) {
     if (head == NULL) {
-        return 0; 
+        return 0;
     }
 
-    unsigned long long result = 1;
+    long long result = 1;
     struct Node* current = head;
 
     while (current != NULL) {
-        result *= (unsigned long long)current->data;
+        /* Possible weaknesses found:
+         * UBSan: signed integer overflow: 4609047196709527572 * 3 cannot be represented in type 'long long' (AFL crash: id:000000,sig:06,src:000001,time:1012,execs:936,op:havoc,rep:4)
+         */
+        result *= current->data;
         current = current->next;
     }
 
-    return (long long)result;
+    return result;
 }
 
-struct Node* createNode(long long data) {
+struct Node* createNode(int data) {
     struct Node* newNode = (struct Node*)malloc(sizeof(struct Node));
     if (newNode == NULL) {
         exit(EXIT_FAILURE);
@@ -49,11 +52,11 @@ void freeList(struct Node* head) {
 }
 
 int main(void) {
-    long long val;
+    int val;
     struct Node* head = NULL;
     struct Node* tail = NULL;
 
-    while (scanf("%lld", &val) == 1) {
+    while (scanf("%d", &val) == 1) {
         struct Node* newNode = createNode(val);
         if (head == NULL) {
             head = newNode;
