+++ collected_code_6/gemini-3-pro/feedback-static+dynamic-r0-t1.0-p1.0/heal_5/to_be_healed/problem-919-s3.c
@@ -6,9 +6,9 @@ struct Node {
     struct Node* next;
 };
  /* Possible weaknesses found:
-  *  test case 0 failed: expected -6, got <no output>
   *  test case 2 failed: expected 18, got <no output>
   *  test case 1 failed: expected 24, got <no output>
+  *  test case 0 failed: expected -6, got <no output>
   */
 
 long long multiplyListItems(struct Node* head) {
@@ -16,18 +16,15 @@ long long multiplyListItems(struct Node* head) {
         return 0; 
     }
 
-    long long result = 1;
+    unsigned long long result = 1;
     struct Node* current = head;
 
     while (current != NULL) {
-        /* Possible weaknesses found:
-         * UBSan: signed integer overflow: -8637036320425427712 * 3 cannot be represented in type 'long long' (AFL crash: id:000000,sig:06,src:000008,time:54055,execs:5246,op:havoc,rep:4)
-         */
-        result *= (long long)(current->data);
+        result *= (unsigned long long)(long long)current->data;
         current = current->next;
     }
 
-    return result;
+    return (long long)result;
 }
 
 struct Node* createNode(int data) {
@@ -68,6 +65,7 @@ int main(void) {
     }
 
     long long product = multiplyListItems(head);
+
     printf("%lld\n", product);
 
     freeList(head);
