+++ collected_code_6/gemini-3-pro/feedback-static+dynamic-r0-t1.0-p1.0/heal_9/healed/problem-919-s3.c
@@ -2,27 +2,27 @@
 #include <stdlib.h>
 
 struct Node {
-    long long data;
+    int data;
     struct Node* next;
 };
 
 long long multiplyListItems(struct Node* head) {
     if (head == NULL) {
-        return 0; 
+        return 0;
     }
 
-    unsigned long long result = 1;
+    long long result = 1;
     struct Node* current = head;
 
     while (current != NULL) {
-        result *= (unsigned long long)current->data;
+        result *= current->data;
         current = current->next;
     }
 
-    return (long long)result;
+    return result;
 }
 
-struct Node* createNode(long long data) {
+struct Node* createNode(int data) {
     struct Node* newNode = (struct Node*)malloc(sizeof(struct Node));
     if (newNode == NULL) {
         exit(EXIT_FAILURE);
@@ -44,11 +44,11 @@ void freeList(struct Node* head) {
 }
 
 int main(void) {
-    long long val;
+    int val;
     struct Node* head = NULL;
     struct Node* tail = NULL;
 
-    while (scanf("%lld", &val) == 1) {
+    while (scanf("%d", &val) == 1) {
         struct Node* newNode = createNode(val);
         if (head == NULL) {
             head = newNode;
