+++ collected_code_6/glm-5.1/feedback-static+dynamic-r0-t1.0-p1.0/heal_3/to_be_healed/problem-442-s3.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 #include <stdlib.h>
  /* Possible weaknesses found:
+  *  test case 1 failed: expected 0.69, got 5
   *  test case 2 failed: expected 0.56, got 3
   *  test case 0 failed: expected 0.54, got 3
-  *  test case 1 failed: expected 0.69, got 5
   */
 
 int positive_count(const int *array, size_t size) {
@@ -20,7 +20,7 @@ double find_ratio(const int *array, size_t size) {
     if (size == 0) {
         return 0.0;
     }
-    return positive_count(array, size) / (double)size;
+    return (double)positive_count(array, size) / (double)size;
 }
 
 int main(void) {
