+++ collected_code_6/qwen-max/feedback-static+dynamic-r0-t1.0-p1.0/heal_2/to_be_healed/problem-422-s2.c
@@ -1,17 +1,17 @@
 #include <stdio.h>
 #include <stdlib.h>
- /* Possible weaknesses found:
-  *  test case 1 failed: expected 12, got <no output>
-  *  test case 2 failed: expected 1, got <no output>
-  *  test case 0 failed: expected 4.5, got <no output>
-  */
+#include <stdint.h>
 
 double average_of_cubes(int n) {
     if (n <= 0) return 0.0;
 
-    long long sum = 0;
+    int64_t sum = 0;
     for (int i = 1; i <= n; ++i) {
-        sum += (long long)i * i * i;
+        /* Possible weaknesses found:
+         * UBSan: signed integer overflow: 9223193340756366400 + 473384830201856 cannot be represented in type 'int64_t' (aka 'long long') (AFL crash: id:000000,sig:06,src:000001,time:117,execs:140,op:havoc,rep:4)
+         * UBSan: signed integer overflow: 4398046511104 * 2097152 cannot be represented in type 'int64_t' (aka 'long long') (AFL crash: id:000000,sig:06,src:000001,time:117,execs:140,op:havoc,rep:4)
+         */
+        sum += (int64_t)i * i * i;
     }
     return (double)sum / n;
 }
