+++ collected_code_6/qwen-max/feedback-static+dynamic-r0-t1.0-p1.0/heal_2/to_be_healed/problem-960-s3.c
@@ -1,12 +1,13 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
-int tile(int n) {
+int64_t tile(int n) {
     if (n == 0 || n == 1) return 1;
-    int a = 1, b = 1, c = 0;
+    int64_t a = 1, b = 1, c = 0;
     for (int i = 2; i <= n; i++) {
         /* Possible weaknesses found:
-         * UBSan: signed integer overflow: 1134903170 + 1836311903 cannot be represented in type 'int' (AFL crash: id:000000,sig:06,src:000002,time:409,execs:199,op:havoc,rep:2)
+         * UBSan: signed integer overflow: 4660046610375530309 + 7540113804746346429 cannot be represented in type 'int64_t' (aka 'long long') (AFL crash: id:000000,sig:06,src:000001,time:308,execs:282,op:havoc,rep:2)
          */
         c = a + b;
         a = b;
@@ -26,7 +27,10 @@ int main() {
         fprintf(stderr, "Number of tiles cannot be negative\n");
         return EXIT_FAILURE;
     }
-    int result = tile(n);
-    printf("The number of ways to tile the floor is: %d\n", result);
+    int64_t result = tile(n);
+    /* Possible weaknesses found:
+     *  %lld in format string (no. 1) requires 'long long' but the argument type is 'signed long'. [invalidPrintfArgType_sint]
+     */
+    printf("The number of ways to tile the floor is: %lld\n", result);
     return EXIT_SUCCESS;
 }
