+++ collected_code_6/qwen-max/feedback-static+dynamic-r0-t1.0-p1.0/heal_3/to_be_healed/problem-529.c
@@ -1,23 +1,27 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <inttypes.h>
  /* Possible weaknesses found:
-  *  test case 0 failed: expected 31, got <no output>
-  *  test case 2 failed: expected 17, got <no output>
-  *  test case 1 failed: expected 5, got <no output>
+  *  test case 0 failed: expected 31, got 53
+  *  test case 2 failed: expected 17, got 22
+  *  test case 1 failed: expected 5, got 4
   */
 
 long long jacobsthal_lucas(long n) {
     if (n == 0) return 2;
     if (n == 1) return 1;
 
-    long long a = 2, b = 1;
+    int64_t a = 2, b = 1;
     for (long i = 2; i <= n; i++) {
-        long long c = a + 2 * b;
+        int64_t c = a + 2 * b;
         a = b;
         b = c;
     }
     return b;
 }
+ /* Possible weaknesses found:
+  * Fuzzing found a crash (signal 6) on input id:000000,sig:06,src:000001,time:101,execs:126,op:havoc,rep:2; likely memory-safety defect
+  */
 
 int main() {
     long n;
@@ -27,7 +31,7 @@ int main() {
         return EXIT_FAILURE;
     }
 
-    long long result = jacobsthal_lucas(n);
-    printf("The %ldth Jacobsthal-Lucas number is: %lld\n", n, result);
+    int64_t result = jacobsthal_lucas(n);
+    printf("The %ldth Jacobsthal-Lucas number is: %" PRId64 "\n", n, result);
     return EXIT_SUCCESS;
 }
